distanceInKm in constructor changed to float

+++ src/main/java/com/codecool/shipping_cost_calculator_service/model/Option.java
@@ -23,7 +23,7 @@ public class Option {
 //    -------------------------------
 
 
-    public Option(String type, String originFound, String destinationFound, int distanceInKm, float timeInHours) {
+    public Option(String type, String originFound, String destinationFound, float distanceInKm, float timeInHours) {
         this.originFound = originFound;
         this.destinationFound = destinationFound;
